CIV-9457 Update from review
